Work

+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: oroy <oroy@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/15 15:32:25 by oroy              #+#    #+#             */
-/*   Updated: 2023/06/22 15:32:27 by oroy             ###   ########.fr       */
+/*   Updated: 2023/06/29 18:50:46 by oroy             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -78,6 +78,6 @@ t_list	*ft_lstlast(t_list *lst);
 t_list	*ft_lst2ndlast(t_list *lst);
 t_list	*ft_lstmap(t_list *lst, void *(*f)(void *), void (*del)(void *));
 t_list	*ft_lstnew(int content, t_list *prev);
-int		ft_lstsize(t_list *lst);
+size_t	ft_lstsize(t_list *lst);
 
 #endif
